Refactor nonce-based auth (#1462)
* Refactor nonce-based auth

We no longer require a wallet address when generating a nonce, and will accept any signed message to prove ownership of a wallet as long as the message contains one of our generated nonces. Nonces are now guaranteed to be one-use only, and they expire after an hour.

* Get rid of unused interface

* Add usersByAddresses query

* Allow optionally importing verified Farcaster addresses via autosocial

* Format GraphQL

+++ service/persist/postgres/user.go
@@ -176,14 +176,16 @@ func (u *UserRepository) createWalletWithTx(ctx context.Context, queries *db.Que
 	walletID := wallet.ID
 
 	if walletID != "" {
-
 		user, err := queries.GetUserByWalletID(ctx, walletID.String())
 		if err != nil && !errors.Is(err, pgx.ErrNoRows) {
 			return "", err
 		}
 
-		// If the wallet belongs to a user, its address can't be used to create a new wallet. Return an error.
 		if user.ID != "" {
+			if user.ID == userID {
+				// Wallet already belongs to the user; do nothing
+				return walletID, nil
+			}
 			if user.Universal {
 				err = queries.DeleteUserByID(ctx, user.ID)
 				if err != nil {
